+++ lib/src/chart/bar_chart/bar_chart_painter.dart
@@ -1,10 +1,9 @@
+import 'package:fl_chart/src/chart/bar_chart/bar_chart_data.dart';
+import 'package:fl_chart/src/chart/base/axis_chart/axis_chart_painter.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 
-import '../bar_chart/bar_chart_data.dart';
-import '../base/fl_axis_chart/fl_axis_chart_painter.dart';
-
-class BarChartPainter extends FlAxisChartPainter {
+class BarChartPainter extends AxisChartPainter {
   final BarChartData data;
 
   Paint barPaint;