+++ CHANGELOG.md
@@ -3,3 +3,8 @@
 ## 0.0.2
 * fixed `minX`, `maxX` functionality on LineChart
 * restricted to access private classes of the library
+
+## 0.0.3
+* renamed `FlChartWidget` to `FlChart` (our main widget) and now you have to import `package:fl_chart/fl_chart.dart` instead of `package:fl_chart/fl_chart_widget.dart`
+* renamed `FlChart*` to `BaseChart*` (parent class of our charts like `PieChart`)
+* renamed `FlAxisChart*` to `AxisChart*`